% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_splicing.R
\name{beta_real}
\alias{beta_real}
\title{Convert K-1 representation of parameters to real}
\usage{
beta_real(r)
}
\arguments{
\item{r}{A parameter object from fitting dm_glm or dm_glm_multi_conc. Must have members beta_raw and beta_scale.}
}
\value{
Coefficient matrix.
}
\description{
Convert K-1 representation of parameters to real
}
