% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_splicing.R
\name{cluster_results_table}
\alias{cluster_results_table}
\title{Parse output of differential_splicing}
\usage{
cluster_results_table(results)
}
\arguments{
\item{results}{From \code{\link{differential_splicing}}}
}
\value{
Data.frame with columns status, log likelihood ratio, degrees of freedom, p-value
}
\description{
Parse output of \code{\link{differential_splicing}} and make a per cluster results table
}
