% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_glm.R
\name{dirichlet_multinomial_anova}
\alias{dirichlet_multinomial_anova}
\title{Dirichlet multinomial GLM likelihood ratio test for a single cluster (single overdispersion parameter)}
\usage{
dirichlet_multinomial_anova(xFull, xNull, y, concShape = 1.0001,
  concRate = 1e-04)
}
\arguments{
\item{xFull}{[samples] x [covariates] matrix for the alternative model}

\item{xNull}{[samples] x [covariates] matrix for the null model}

\item{y}{[samples] x [introns] matrix of intron usage counts}

\item{concShape}{Gamma shape parameter for concentration parameter}

\item{concRate}{Gamma rate parameter for concentration parameter}
}
\description{
We recommend using \code{\link{dirichlet_multinomial_glm_anova_mc}} instead.
}
