% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_glm_multi_conc.R
\name{dirichlet_multinomial_anova_mc}
\alias{dirichlet_multinomial_anova_mc}
\title{Dirichlet multinomial GLM likelihood ratio test for a single cluster}
\usage{
dirichlet_multinomial_anova_mc(xFull, xNull, y, concShape = 1.0001,
  concRate = 1e-04, fit_null = NULL)
}
\arguments{
\item{xFull}{[samples] x [covariates] matrix for the alternative model}

\item{xNull}{[samples] x [covariates] matrix for the null model}

\item{y}{[samples] x [introns] matrix of intron usage counts}

\item{concShape}{Gamma shape parameter for concentration parameter}

\item{fit_null}{Optionally the fitted null model (used in \code{\link{splicing_qtl}} to save repeatedly fitting the null for each cis-SNP)}

\item{concShape}{Gamma rate parameter for concentration parameter}
}
\description{
Dirichlet multinomial GLM likelihood ratio test for a single cluster
}

