% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_glm.R
\name{dirichlet_multinomial_glm}
\alias{dirichlet_multinomial_glm}
\title{Dirichlet multinomial GLM (single overdispersion parameter)}
\usage{
dirichlet_multinomial_glm(x, y, concShape = 1.0001, concRate = 1e-04)
}
\arguments{
\item{x}{[samples] x [covariates] matrix}

\item{y}{[samples] x [introns] matrix of intron usage counts}

\item{concShape}{Gamma shape parameter for concentration parameter}

\item{concRate}{Gamma rate parameter for concentration parameter}
}
\description{
We recommend using \code{\link{dirichlet_multinomial_anova_mc}} instead.
}
