% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_glm_multi_conc.R
\name{dirichlet_multinomial_glm_mc}
\alias{dirichlet_multinomial_glm_mc}
\title{Dirichlet multinomial GLM}
\usage{
dirichlet_multinomial_glm_mc(x, y, concShape = 1.0001, concRate = 1e-04)
}
\arguments{
\item{x}{[samples] x [covariates] matrix}

\item{y}{[samples] x [introns] matrix of intron usage counts}

\item{concShape}{Gamma shape parameter for concentration parameter}

\item{concShape}{Gamma rate parameter for concentration parameter}
}
\description{
Dirichlet multinomial GLM
}

