% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_intron_meta}
\alias{get_intron_meta}
\title{Make a data.frame of meta data about the introns}
\usage{
get_intron_meta(introns)
}
\arguments{
\item{introns}{Names of the introns}
}
\value{
Data.frame with chr, start, end, cluster id and "middle"
}
\description{
Make a data.frame of meta data about the introns
}
