% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_splicing.R
\name{leaf_cutter_effect_sizes}
\alias{leaf_cutter_effect_sizes}
\title{Per intron effect sizes}
\usage{
leaf_cutter_effect_sizes(results)
}
\arguments{
\item{results}{From \code{\link{differential_splicing}}}
}
\value{
Numeric vector of effect sizes for every tested intron.
}
\description{
Per intron effect sizes
}

