% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_splicing.R
\name{leafcutter_status}
\alias{leafcutter_status}
\title{Differential splicing statuses}
\usage{
leafcutter_status(results)
}
\arguments{
\item{results}{From \code{\link{differential_splicing}}}
}
\value{
Character vector with statuses
}
\description{
Get a character vector saying what happened for each tested cluster.
}

