% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cluster_plot.R
\name{make_cluster_plot}
\alias{make_cluster_plot}
\title{Make cluster level plot}
\usage{
make_cluster_plot(cluster_to_plot, main_title = NA, exons_table = NULL,
  meta = NULL, cluster_ids = NULL, counts = NULL, introns = NULL,
  snp_pos = NA)
}
\description{
Make cluster level plot
}
