% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/junction_plot.R
\name{make_differential_splicing_plot}
\alias{make_differential_splicing_plot}
\title{Make sashimi-esque plot with ggplot2}
\usage{
make_differential_splicing_plot(y, x, len = 500,
  length_transform = function(g) log(g + 1), main_title = NA,
  snp_pos = NA, summary_func = colMeans, legend_title = "Mean count")
}
\arguments{
\item{y}{[samples] x [introns] matrix of counts}

\item{x}{[samples] vector of group membership. Can be numeric, factor or character.}

\item{len}{Number of segments each curve is constructed from, controls smoothness.}

\item{length_transform}{Function controlling how true genomic space is mapped to the plot for improved visability. Use the identity function (i.e. function(g) g}

\item{main_title}{Title}

\item{snp_pos}{An optional list of SNP positions to mark}

\item{summary_func}{Function to summarize counts within groups: usually colMeans or colSums}

\item{Legend}{title Match this to summary_func}
}
\description{
Shows only junction reads and can optionally show splicing variation across groups.
}

