% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_clusters_to_genes}
\alias{map_clusters_to_genes}
\title{Work out which gene each cluster belongs to. Note the chromosome names used in the two inputs must match.}
\usage{
map_clusters_to_genes(intron_meta, exons_table)
}
\arguments{
\item{intron_meta}{Data frame describing the introns, usually from get_intron_meta}

\item{exons_table}{Table of exons, see e.g. /data/gencode19_exons.txt.gz}
}
\value{
Data.frame with cluster ids and genes separated by commas
}
\description{
Work out which gene each cluster belongs to. Note the chromosome names used in the two inputs must match.
}
