% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splicing_qtl.R
\name{splicing_qtl}
\alias{splicing_qtl}
\title{Perform splicing QTL analysis}
\usage{
splicing_qtl(counts, geno, geno_meta, pcs = matrix(0, ncol(counts), 0),
  permute = F, snps_within = 10000, min_samples_per_intron = 5,
  min_coverage = 20, min_samples_per_group = 8, timeout = 10, debug = F,
  outlier_threshold = 1e-30, robust = T, ...)
}
\arguments{
\item{counts}{An [introns] x [samples] matrix of counts. The rownames must be of the form chr:start:end:cluid. If the counts file comes from the leafcutter clustering code this should be the case already.}

\item{geno}{A [SNPs] x [samples] numeric matrix of the genotypes, typically encoded as 0,1,2, although in principle scaling shouldn't matter.}

\item{geno_meta}{SNP metadata, as a data.frame. Rows correspond to SNPs, must have a CHROM (with values e.g. chr15) and POS (position) column.}

\item{pcs}{An optional [confounders] x [samples] matrix of technical confounders/PCs.}

\item{permute}{Whether to permute the genotype.}

\item{snps_within}{Window from center of cluster in which to test SNPs.}

\item{min_samples_per_intron}{Ignore introns used (i.e. at least one supporting read) in fewer than n samples}

\item{min_coverage}{Require min_samples_per_group samples in each group to have at least this many reads}

\item{min_samples_per_group}{Require this many samples in each group to have at least min_coverage reads}

\item{timeout}{Maximum time (in seconds) allowed for a single optimization run}

\item{debug}{Turn on to see output from rstan.}

\item{max_cluster_size}{Don't test clusters with more introns than this}
}
\value{
A per cluster list of results. For each cluster this is a list over tested SNPs. SNPs that were not tested will be represented by a string saying why.
}
\description{
Parallelization across tested clusters is achieved using foreach/doMC, so the number of threads that will be used is determined by the cores argument passed to registerDoMC.
}

