% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_glm.R
\name{betaBinomialGLM}
\alias{betaBinomialGLM}
\title{Beta binomial GLM likelihood ratio testing}
\usage{
betaBinomialGLM(ys, ns, xFull, xNull, concShape = 1.0001, concRate = 1e-04,
  ...)
}
\arguments{
\item{ys}{numerator counts}

\item{ns}{denominator counts}

\item{xFull}{matrix of covariates for full model. First column must be ones.}

\item{xNull}{matrix of covariates for null model. First column must be ones.}

\item{concShape}{Gamma shape parameter for concentration parameter}

\item{concRate}{Gamma rate parameter for concentration parameter}

\item{...}{will be passed on the rstan::optimizing, so can be used for example to set the algorithm used (default is LBFGS).}
}
\description{
We recommend using \code{\link{dirichlet_multinomial_glm_anova}} instead, but this is here for comparison.
}
