% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_glm_multi_conc.R
\name{dirichlet_multinomial_anova_mc}
\alias{dirichlet_multinomial_anova_mc}
\title{Dirichlet multinomial GLM likelihood ratio test for a single cluster}
\usage{
dirichlet_multinomial_anova_mc(xFull, xNull, y, concShape = 1.0001,
  concRate = 1e-04, robust = T, outlier_prior_a = 1.01,
  outlier_prior_b = 100, fit_null = NULL, debug = F, ...)
}
\arguments{
\item{xFull}{[samples] x [covariates] matrix for the alternative model}

\item{xNull}{[samples] x [covariates] matrix for the null model}

\item{y}{[samples] x [introns] matrix of intron usage counts}

\item{concShape}{Gamma shape parameter for concentration parameter}

\item{concRate}{Gamma rate parameter for concentration parameter}

\item{robust}{Whether to include an outlier model (i.e. use dm_glm_multi_conc_robust rather than dm_glm_multi_conc)}

\item{outlier_prior_a}{Only used for robust model. The outlier probability outlier_prob ~ Beta(outlier_prior_a,outlier_prior_b)}

\item{outlier_prior_b}{Only used for robust model. The outlier probability outlier_prob ~ Beta(outlier_prior_a,outlier_prior_b)}

\item{fit_null}{Optionally cache the fitted null model to save repeatedly fitting the null for each cis-SNP when sQTL mapping)}

\item{debug}{Whether to give verbose output from rstan.}

\item{...}{will be passed on the rstan::optimizing, so can be used for example to set the algorithm used (default is LBFGS).}
}
\description{
Dirichlet multinomial GLM likelihood ratio test for a single cluster
}
