% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiqq}
\alias{multiqq}
\title{Plot multiple qq plot for p-values}
\usage{
multiqq(pvalues)
}
\arguments{
\item{pvalues}{A named list of numeric vectors of p-values.}
}
\value{
a ggplot
}
\description{
Plot multiple qq plot for p-values
}
